Day-6: Add a program which gives Average of an array

+++ Day-6/AvgOfArr.java
@@ -12,12 +12,12 @@ public class AvgOfArr {
             System.out.print("Enter element " + (i + 1)+ " : ");
             arr[i] = sc.nextInt();
         }
-        int sum = 0;
+        double sum = 0;
         for (int i = 0; i < arr.length; i++) {
             sum += arr[i];
         }
         System.out.println("The original array is : "+ Arrays.toString(arr));
-        int avg = sum/n;
+        double avg = sum/n;
         System.out.println("The avg of the array is : " + avg);
         sc.close();
     }
